Disable new on-by-default ImageMagick heif option.
libheif needs libde265, which needs SDL, which pulls in a bunch of
X11 stuff, which I don't really want but could maybe live with, if
it didn't break my build. fixesproto apparently wants to install
include/X11/extensions/xfixesproto.h but it can't because xorgproto
already did.

I'll untangle this later if and when I need HEIF in ImageMagick.